fix sdr_transceiver_wspr/app/stop.sh

+++ projects/sdr_transceiver_wspr/app/stop.sh
@@ -1,7 +1,8 @@
 #! /bin/sh
 
-rm -f /etc/periodic/wspr
-
-service dcron restart
-
-killall -q decode-wspr.sh write-c2-files wsprd
+if test -e /etc/periodic/wspr
+then
+  rm -f /etc/periodic/wspr
+  service dcron restart
+  killall -q decode-wspr.sh write-c2-files wsprd
+fi
